move full install to new module
