+++ npe2/_inspection/_fetch.py
@@ -3,10 +3,7 @@ from __future__ import annotations
 import json
 import os
 import re
-import site
-import sys
 import tempfile
-import warnings
 from concurrent.futures import ProcessPoolExecutor
 from contextlib import contextmanager
 from functools import lru_cache
@@ -19,13 +16,9 @@ from urllib import error, parse, request
 from urllib.request import urlopen
 from zipfile import ZipFile
 
-from build.env import IsolatedEnvBuilder
-
 from npe2.manifest import PackageMetadata
 
 if TYPE_CHECKING:
-    import build.env
-
     from npe2.manifest import PluginManifest
 
 
@@ -36,7 +29,6 @@ NPE2_ENTRY_POINT = "napari.manifest"
 __all__ = [
     "fetch_manifest",
     "get_pypi_url",
-    "isolated_plugin_env",
     "get_hub_plugin",
     "get_hub_plugins",
     "get_pypi_plugins",
@@ -273,26 +265,6 @@ def _tmp_pypi_sdist_download(
 
 
 @lru_cache
-def get_hub_plugins() -> Dict[str, str]:
-    """Return {name: latest_version} for all plugins on the hub."""
-    with urlopen("https://api.napari-hub.org/plugins") as r:
-        return json.load(r)
-
-
-@lru_cache
-def get_hub_plugin(plugin_name: str) -> Dict[str, Any]:
-    """Return hub information for a specific plugin."""
-    with urlopen(f"https://api.napari-hub.org/plugins/{plugin_name}") as r:
-        return json.load(r)
-
-
-def get_pypi_plugins() -> Dict[str, str]:
-    """Return {name: latest_version} for all plugins found on pypi."""
-    NAPARI_CLASSIFIER = "Framework :: napari"
-    return _get_packages_by_classifier(NAPARI_CLASSIFIER)
-
-
-@lru_cache
 def _get_packages_by_classifier(classifier: str) -> Dict[str, str]:
     """Search for packages declaring ``classifier`` on PyPI.
 
@@ -319,6 +291,26 @@ def _get_packages_by_classifier(classifier: str) -> Dict[str, str]:
             break
 
     return dict(sorted(packages.items()))
+
+
+def get_pypi_plugins() -> Dict[str, str]:
+    """Return {name: latest_version} for all plugins found on pypi."""
+    NAPARI_CLASSIFIER = "Framework :: napari"
+    return _get_packages_by_classifier(NAPARI_CLASSIFIER)
+
+
+@lru_cache
+def get_hub_plugins() -> Dict[str, str]:
+    """Return {name: latest_version} for all plugins on the hub."""
+    with urlopen("https://api.napari-hub.org/plugins") as r:
+        return json.load(r)
+
+
+@lru_cache
+def get_hub_plugin(plugin_name: str) -> Dict[str, Any]:
+    """Return hub information for a specific plugin."""
+    with urlopen(f"https://api.napari-hub.org/plugins/{plugin_name}") as r:
+        return json.load(r)
 
 
 def _try_fetch_and_write_manifest(args: Tuple[str, str, Path]):
@@ -350,93 +342,3 @@ def fetch_all_manifests(dest: str = "manifests"):
         errors = list(executor.map(_try_fetch_and_write_manifest, args))
     _errors = {tup[0]: tup[1] for tup in errors if tup}
     (_dest / "errors.json").write_text(json.dumps(_errors, indent=2))
-
-
-@contextmanager
-def isolated_plugin_env(
-    package: str,
-    version: Optional[str] = None,
-    validate_npe1_imports: bool = True,
-    install_napari_if_necessary: bool = True,
-) -> Iterator[build.env.IsolatedEnv]:
-    """Isolated env context with a plugin installed.
-
-    The site-packages folder of the env is added to sys.path within the context.
-
-    Parameters
-    ----------
-    package : str
-        package name
-    version : Optional[str]
-        package version, by default, latest version.
-    validate_npe1_imports: bool
-        Whether to try to import an npe1 plugin's entry points. by default True.
-    install_napari_if_necessary: bool
-        If `validate_npe1_imports` is True, whether to install napari if the import
-        fails.  (It's not uncommon for plugins to fail to specify napari as a
-        dependency.  Othertimes, they simply need a qt backend.).  by default True.
-
-    Yields
-    ------
-    build.env.IsolatedEnv
-        env object that has an `install` method.
-    """
-    with IsolatedEnvBuilder() as env:
-        # install the package
-        pkg = f"{package}=={version}" if version else package
-        logger.debug(f"installing {pkg} into virtual env")
-        env.install([pkg])
-
-        # temporarily add env site packages to path
-        prefixes = [getattr(env, "path")]  # noqa
-        if not (site_pkgs := site.getsitepackages(prefixes=prefixes)):
-            raise ValueError("No site-packages found")  # pragma: no cover
-        sys.path.insert(0, site_pkgs[0])
-        try:
-            if validate_npe1_imports:
-                # try to import the plugin's entry points
-                dist = metadata.distribution(package)
-                ep_groups = {ep.group for ep in dist.entry_points}
-                if NPE1_ENTRY_POINT in ep_groups and NPE2_ENTRY_POINT not in ep_groups:
-                    try:
-                        _get_loaded_mf_or_die(package)
-                    except Exception:  # pragma: no cover
-                        # if loading contributions fails, it can very often be fixed
-                        # by installing `napari[all]` into the environment
-                        if install_napari_if_necessary:
-                            env.install(["napari[all]"])
-                            # force reloading of qtpy
-                            sys.modules.pop("qtpy", None)
-                            _get_loaded_mf_or_die(package)
-                        else:
-                            raise
-            yield env
-        finally:
-            # cleanup sys.path
-            sys.path.pop(0)
-
-
-def _get_loaded_mf_or_die(package: str) -> PluginManifest:
-    """Return a fully loaded (if npe1) manifest, or raise an exception."""
-    from npe2 import PluginManifest
-    from npe2.manifest._npe1_adapter import NPE1Adapter
-
-    mf = PluginManifest.from_distribution(package)
-    if isinstance(mf, NPE1Adapter):
-        with warnings.catch_warnings():
-            warnings.filterwarnings("error", message="Error importing contributions")
-            warnings.filterwarnings("error", message="Failed to convert")
-            warnings.filterwarnings("ignore", message="Found a multi-layer writer")
-            mf._load_contributions(save=False)
-    return mf
-
-
-def fetch_manifest_with_full_install(
-    package: str, version: Optional[str] = None
-) -> PluginManifest:
-    """Fetch manifest for plugin by installing into an isolated environment."""
-    # create an isolated env in which to install npe1 plugin
-    with isolated_plugin_env(
-        package, version, validate_npe1_imports=True, install_napari_if_necessary=True
-    ):
-        return _get_loaded_mf_or_die(package)
